+++ src/main/java/kr/inuappcenterportal/inuportal/domain/firebase/dto/TrackedNotificationDispatch.java
@@ -1,0 +1,16 @@
+package kr.inuappcenterportal.inuportal.domain.firebase.dto;
+
+import kr.inuappcenterportal.inuportal.domain.firebase.enums.FcmMessageType;
+
+import java.util.Map;
+
+public record TrackedNotificationDispatch(
+        Long fcmMessageId,
+        Map<String, Long> tokenAndMemberId,
+        String title,
+        String body,
+        FcmMessageType type,
+        Long targetId,
+        String path
+) {
+}
